Add documentation on 'getting up and running'
Some hopefully usefully instructions to help people get up and running.

Re: Issue #34 On my machine I didn't need to install memcached to successfully install the local requirements.

Re: Issue #39 I've linked to this issue for now. It would be good to remove the need for those steps and then this link can be removed.